Run unit tests in a docker container
Some of the routing-release components require tests to be run on a
Linux machine, so we're providing a script `run-unit-tests-in-docker`
for that purpose. This script will actually call the `bin/test` script
of each component so this commit may cause failures as due to outdated
`bin/test` scripts. If tests break, we should fix them accordingly.

[finishes #161301193]

Co-authored-by: Utako Ueda <uueda@pivotal.io>