Change config for add_headers_if_not_present
We want to change the format of the config to rewrite the HTTP
response to add headers if the are not present, using
the more semanting `responses.add_headers_if_not_present`
instead of inject.

The fature covers be able to add headers in any response from the
gorouter. The header would be added only if the backend
does not return headers with the same value.

One example use case can be add a HSTS header,
Strict-Transport-Security into any response from the gorouter[1]

We allow some new configuration as in the router as:

    router:
      http_rewrite:
        responses:
          add_headers_if_not_present:
          - name: Strict-Transport-Security
            value: max-age=31536000; includeSubDomains; preload
          - name: X-A-Header
            value: the-value

See:
 - https://github.com/cloudfoundry/gorouter/pull/232
 - https://github.com/cloudfoundry/gorouter/pull/234 (config struct change)