Allow inject custom headers into gorouter response
We want to be able to inject the heades in any response from the
go router. The header would be injected only if the backend
does not return those headers.

One example use case can be add a HSTS header,
Strict-Transport-Security into any response from the gorouter[1]

We allow some new configuration as in the router as:

    http_rewrite:
      inject_response_headers:
      - name: X-A-Header
      value: the-value
      - name: X-A-Header
      value: other-value
      - name: X-Other-Header
      value: some value

See https://github.com/cloudfoundry/gorouter/pull/232