Allow remove headers from response
We want to be able to define a list of headers to be stripped
form the HTTP responses by the gorouter.

We add a new entry under router.http_rewirte. No need to modify
the templates, as they already dump the structure as json and
it matches the config format in gorouter.

See: https://github.com/cloudfoundry/gorouter/pull/234