ci: reorder jubilee tasks
We discovered that we can't check the results of the jubilee tests after
the deployment is deleted. We believe that by moving the delete to the
last step that if something does fail, we will feel the pain from it
breaking and know that we need to look at it.

Co-authored-by: Clay Kauzlaric <ckauzlaric@pivotal.io>