ci: add nats bosh-dns-aliases to zd deployment
* zero-downtime-release (used in jubilee) is deployed with a route
registrar to register the zero downtime backend routes with the
gorouter. Route Registrar consumes nats from the cf deployed alongside
the zero-downtime deployment. However, with the change to nats recently
to use a hostname instead of an ip, route registrar in the zero-downtime
deployment tries to connect to nats using `nats.service.cf.internal`,
but there is no bosh-dns-alias for that in the zero-downtime deployment. This commit
adds that alias.
* failing jobs: https://networking.ci.cf-app.com/teams/ga/pipelines/routing/jobs/jubilee-no-downtime/builds/264#L5eb1b6c7:102