ci: wait longer before run indicator-protocol-test
A lot of runs of this job are failing for what appears to be a timing
issue. When we intercept the container after a failed test and attempt
the indicator-protocol-test again, we see that all indicators are
present.

We are hoping to add a little more delay to ensure that indicators are
present and make this more configurable in the future if necessary.