ci: retry indicator-protocol-test up to 3 times
Similar to the concept in d7cad9c98113c95d60b67d91db0ee94d2d7ccc6, we
notice that running the tests immediately afterwards in the intercepted
container passes without issue. In this commit, we add retries to
attempt the indicator-protocol-test up to three times in order to
hopefully assuage this flake.

Also defaults `reconfigure` script to set `routing` pipeline if no param
is provided.