Add router_ssl.ca to trusted certs in ci-lite
In bosh-lite, there is no LB, so gorouter terminates SSL. In order for
the route-service tests to pass, gorouter needs to trust the certificate
authority used to sign the certificate it presents when it makes calls
back to itself.

We chose to add a separate opsfile to make it clear that this is a
special case for bosh lite.

Signed-off-by: David McClure <dmcclure@vmware.com>