DO NOT ADD THIS TO PR
these are local changes necessary to get everythign to work

set dns alias

this allows a fqdn to be set so that we may access the routing-api
with a url

[#175661519]

add dns alias

this alias is required to resolve the .internal fqdn
of the sni routing api

[#175645646]

use s3

we expect the blob store to be s3 for our precompilation work

this should fix our build

put blobs in our s3 bucket

we need write access to these blob locations
we also need read access

just put them in out blobstore.

fix create final release

because we are using our own bucket so that
we can have write acess, we need to recreate
the vendored packages and dependent blobs as
well as all references to the old blob and package
ids that bosh has held.

[]