Configure Gorouter to talk to NATS over mTLS
Until now, Gorouter has talked to NATS over an unencrypted connection.
Authentication was with a password in plaintext. We want to encrypt
and authenticate this connection. For that purpose we're using Mutual
TLS, the same as route-emitter and route-registrar use with NATS.

This commit passes Gorouter the properties it needs from BOSH. The YAML
used for Gorouter's NATS Configuration changes quite a bit as part of this,
so that we don't have to include N copies of the TLS certificates. But the
only new limitation introduced is that all NATS servers have to use the same
username and password.