Be able to use hostname TLS certs for NATS
The TLS certificate presented by NATS is issued for its DNS name, and
not for the individual NATS server IP addresses. Gorouter was being
configured to connect to IP addresses, not the DNS name, and so the
TLS connection failed.

This code is borrowed from route-emitter where someone already had
to solve this [1].

[1] https://github.com/cloudfoundry/diego-release/commit/44260f2401c346dd2d5f73e9587fd88c2b86007e