Bring NATS TLS Opsfile up to date
Since this file was created, route_registrar and NATS are now using TLS in cf-deployment.
Therefore most of this Opsfile is now unnecessary. The nats-smoke-test is retained because
it does not appear to be part of cf-deployment.

This commit forms part of a PR in which we're making Gorouter and NATS talk via TLS, so
the Opsfile now configures that.