Remove reliance on bosh link named 'nats'
We plan to remove plaintext NATS, and enforce that everything uses TLS NATS.
The easiest way to do this is to remove the 'nats' BOSH link, and only have
the 'nats-tls' BOSH link.

This commit changes a file that required the 'nats' BOSH link to exist.
The file will now happily use the 'nats-tls' BOSH link instead.