Raise an error during template generation if tls_pem certs do not have SANs
When gorouter is built with golang 1.17, the `GODEBUG=x509ignoreCN=0`
flag will no longer function, and any deployments using certs without
SANs will fail. This adds a check to fail a deployment at
template-rendering time, to make operators aware of any invalid
certificates. If encountered, `golang.x509ignoreCN` can be set
in the manifest as a stopgap while the cert is recreated with a SAN.

This also changes the default behavior of `golang.x509ignoreCN` to `false`,
to ensure people opt-in and are aware of any invalid certs.

[#175826416](https://www.pivotaltracker.com/story/show/175826416)