Use stndrd imgs in template/unit/integration tests
Modifies the scripts to be able to make use of the standard images:
- cloudfoundry/tas-runtime-build
- cloudfoundry/tas-runtime-mysql-5.7
- cloudfoundry/tas-runtime-mysql-8.0
- cloudfoundry/tas-runtime-postgres

Submodule src/code.cloudfoundry.org/cf-tcp-router 6b149c7f..ecebe81f:
  > Genericize locket setup for postgres too (#13)
Submodule src/code.cloudfoundry.org/routing-api 2bf56685..629f7a89:
  > Genericize test setup for postgres too (#32)

The bumps to `cf-tcp-router` and `routing-api` modify the tests to run
with only one of the sql databases, instead of the assumed two of both
mysql and postgres present in the old image. Now when the
units/integration tests run, all of the components will make use of
mysql when running in a mysql container, and vice versa for postgres.

This also breaks the template-tests out of being run as part of the unit
tests. We already had a script to only run the template-tests:
`scripts/template-tests`, which is easily executed from a docker
container in the other script `scripts/run-template-tests-in-docker`.

[#182691815](https://www.pivotaltracker.com/story/show/182691815)