feat(gorouter): provide option to log more details (#323)
This commit bumps gorouter to include the new feature flag that allows
logging of additional fields in the access log. Those fields provide
additional information on the attempts that were performed and the
timings of the request. The feature flag is exposed in the gorouter job
as `router.enable_log_attempts_details` and defaults to false.