update reserved ports
We audited the ports again and updated the documentation with all the
interesting things we learned.

Found that system-metrics-agent's debug port was not being reserved on
gorouter. Added that to the reserved port list.

Claiming 9100 for otel-collector, which had been reserved for
system-metrics-agent though it was using 53035 in every install we know
of. system-metrics-agent default has been updated to match as well.

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>