Update rubocop config + update bosh template spec tests (#394)
* Update routing's rubocop configuration

I tried to run rubocop the other day and it was an absolute mess. This
PR updates the rubocop config in routing-release and updates all of the
bosh template spec tests appropriately.

The update strategy is based on what exists in CAPI and this PR
specifically:
- https://github.com/cloudfoundry/cloud_controller_ng/pull/3448

* Enable `rubocop-rspec` and regen rubocop_todo.yml

We'll get this these... eventually... maybe.