add additional unsafe params
this allows for more unsafe params to be set.

this is to allow postgres to invoke and access
data persistent disks which are necessary in order
to query the database to determine if a node is
primary during an HA configuration.

allows for privlaged and additional volumes
to be mounted