ci: make docker scripts podman-compatible
Podman is mostly compatible with the docker CLI except for a few small
differences. This commit alignes the docker scripts so that it is
compatible with both.

By default podman requires fully qualified image names and the `rm`
command does not ignore errors when removing a containter that does not
exist so we simply `|| true` after it.