test: do not exhaust cpu
By default tests want to use all available CPUs to execute faster.
In this case however tests do depend on OS doing some background operations,
such as maintenance of the connection status and sockets. The tests have
quite tight timings and there are race conditions that break the tests.

Leave one thread unutilised by tests, so that OS can complete its tasks
promptly, not impacting the tests.

In practice this thread has ca 20-50% cpu load, so it is not that much
cpu that remains unused in the end. Thus overall, tests are just a bit slower.